chore: refactor internal/controller/lifecycle/ pkg contents to controllers/apps/
