+++ .github/utils/utils.sh
@@ -23,6 +23,7 @@ Usage: $(basename "$0") <options>
                                 11) release message
                                 12) send message
                                 13) patch release notes
+                                14) ignore cover pkgs
     -tn, --tag-name           Release tag name
     -gr, --github-repo        Github Repo
     -gt, --github-token       Github token
@@ -33,6 +34,7 @@ Usage: $(basename "$0") <options>
     -tt, --trigger-type       The trigger type (e.g. release/package)
     -ru, --run-url            The run url
     -fl, --file               The release notes file
+    -ip, --ignore-pkgs        The ignore cover pkgs
 EOF
 }
 
@@ -53,6 +55,7 @@ main() {
     local RELEASE_VERSION=""
     local RUN_URL=""
     local FILE=""
+    local IGNORE_PKGS=""
 
     parse_command_line "$@"
 
@@ -95,6 +98,9 @@ main() {
         ;;
         13)
             patch_release_notes
+        ;;
+        14)
+            ignore_cover_pkgs
         ;;
         *)
             show_help
@@ -173,6 +179,12 @@ parse_command_line() {
             -fl|--file)
                 if [[ -n "${2:-}" ]]; then
                     FILE="$2"
+                    shift
+                fi
+                ;;
+            -ip|--ignore-pkgs)
+                if [[ -n "${2:-}" ]]; then
+                    IGNORE_PKGS="$2"
                     shift
                 fi
                 ;;
@@ -413,6 +425,23 @@ patch_release_notes() {
            -X PATCH \
     $GITHUB_API/repos/$GITHUB_REPO/releases/$release_id \
     -d '{"body":"'"$release_note"'"}'
+}
+
+ignore_cover_pkgs() {
+    ignore_pkgs=$(echo "$IGNORE_PKGS" | sed 's/|/ /g')
+    while read line; do
+      ignore=false
+      for pkgs in $(echo "$ignore_pkgs"); do
+          if [[ "$line" == *"$LATEST_REPO/$pkgs"* ]]; then
+              ignore=true
+              break
+          fi
+      done
+      if [[ $ignore == true ]]; then
+          continue
+      fi
+      echo $line >> cover_new.out
+    done < ${FILE}
 }
 
 main "$@"
