+++ controllers/apps/transformer_assure_meta.go
@@ -17,7 +17,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package lifecycle
+package apps
 
 import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
